moved ideas xlsx uploader to separate import tab (thanks Edwin)
