+++ front/app/containers/Admin/ideas/index.tsx
@@ -22,6 +22,11 @@ const IdeasPage = ({ intl: { formatMessage } }: InjectedIntlProps) => {
       name: 'manage',
       url: '/admin/ideas',
     },
+    {
+      label: 'Import',
+      name: 'manage',
+      url: '/admin/ideas/import',
+    },
   ]);
 
   const resource = {
